move MIDICCParameter to the Parameters library (where it belongs)

+++ include/midi/midi_mapper_matrix_types.h
@@ -1,8 +1,8 @@
 #ifndef MATRIX_MAPPER_TYPES__INCLUDED
 #define MATRIX_MAPPER_TYPES__INCLUDED
 
-typedef int8_t source_id_t;
-typedef int8_t target_id_t;
+typedef int source_id_t;
+typedef int target_id_t;
 
 class MIDIMatrixManager;
